build and date

+++ build/jsroot.js
@@ -11,7 +11,7 @@ let version_id = "dev";
 
 /** @summary version date
   * @desc Release date in format day/month/year like "19/11/2021" */
-let version_date = "15/06/2022";
+let version_date = "20/06/2022";
 
 /** @summary version id and date
   * @desc Produced by concatenation of {@link version_id} and {@link version_date}
@@ -55,11 +55,10 @@ function setBatchMode(on) { batch_mode = !!on; }
 /** @summary Indicates if running inside Node.js */
 function isNodeJs() { return nodejs; }
 
-
 let node_atob, node_xhr2;
 
 ///_begin_exclude_in_qt5web_
-
+if (process?.env?.APP_ENV !== 'browser') ;
 ///_end_exclude_in_qt5web_
 
 let browser$1 = { isOpera: false, isFirefox: true, isSafari: false, isChrome: false, isWin: false, touches: false  };
@@ -384,15 +383,19 @@ function getDocument() {
   * @private */
 function injectCode(code) {
    if (nodejs) {
-      let name, fs;
-      return Promise.resolve().then(function () { return _rollup_plugin_ignore_empty_module_placeholder$1; }).then(tmp => {
-         name = tmp.tmpNameSync() + ".js";
-         return Promise.resolve().then(function () { return _rollup_plugin_ignore_empty_module_placeholder$1; });
-      }).then(_fs => {
-         fs = _fs;
-         fs.writeFileSync(name, code);
-         return import("file://" + name);
-      }).finally(() => fs.unlinkSync(name));
+      if (process?.env?.APP_ENV !== 'browser') {
+         let name, fs;
+         return Promise.resolve().then(function () { return _rollup_plugin_ignore_empty_module_placeholder$1; }).then(tmp => {
+            name = tmp.tmpNameSync() + ".js";
+            return Promise.resolve().then(function () { return _rollup_plugin_ignore_empty_module_placeholder$1; });
+         }).then(_fs => {
+            fs = _fs;
+            fs.writeFileSync(name, code);
+            return import("file://" + name);
+         }).finally(() => fs.unlinkSync(name));
+      } else {
+         return Promise.resolve(true); // dummy for webpack
+      }
    }
    if (typeof document !== 'undefined') {
 
@@ -446,12 +449,16 @@ function loadScript(url) {
    let element, isstyle = url.indexOf(".css") > 0;
 
    if (nodejs) {
-      if (isstyle)
-         return Promise.resolve(null);
-      if ((url.indexOf("http:") == 0) || (url.indexOf("https:") == 0))
-         return httpRequest(url, "text").then(code => injectCode(code));
+      if (process.env.APP_ENV !== 'browser') {
+         if (isstyle)
+            return Promise.resolve(null);
+         if ((url.indexOf("http:") == 0) || (url.indexOf("https:") == 0))
+            return httpRequest(url, "text").then(code => injectCode(code));
 
-      return import(url);
+         return import(url);
+      } else {
+         return Promise.resolve(null);
+      }
    }
 
    const match_url = src => {
@@ -45186,7 +45193,7 @@ function assign3DHandler(painter) {
 let node_canvas$1, node_gl;
 
 ///_begin_exclude_in_qt5web_
-
+if (process?.env?.APP_ENV !== 'browser') ;
 ///_end_exclude_in_qt5web_
 
 
@@ -76462,7 +76469,11 @@ class HierarchyPainter extends BasePainter {
          case "draw_tree": return Promise.resolve().then(function () { return TTree; });
          case "hierarchy": return Promise.resolve({ HierarchyPainter, markAsStreamerInfo });
       }
-      return import(module);
+     if (process?.env?.APP_ENV !== 'browser') {
+         return import(module);
+     } else {
+         return Promise.resolve(true);
+     }
    }
 
    /** @summary method used to request object from the http server
@@ -92602,7 +92613,7 @@ TArrowPainter: TArrowPainter
 let node_canvas, btoa_func = globalThis?.btoa;
 
 ///_begin_exclude_in_qt5web_
-
+if (process?.env?.APP_ENV !== 'browser') ;
 ///_end_exclude_in_qt5web_
